updated libft

+++ cursus/libft/testfiles/t_bzero.c
@@ -4,17 +4,17 @@ void	*ft_bzero(void *s, size_t n);
 
 int	main(void)
 {
-	int		i[] = {4, 2, 5, 6, 7, 8};
+	float	i[] = {4, 2, 5, 6, 7, 8};
 	size_t	scan = 0;
 
 	
 	printf("i:\t\t\t");
 	while (scan < 6)
-		printf("%d ", i[scan++]);
+		printf("%f ", i[scan++]);
 	ft_bzero(i, 3);
 	printf("\nft_bzero (s, 3):\t");
 	scan = 0;
 	while (scan < 6)
-		printf("%d ", i[scan++]);
+		printf("%f ", i[scan++]);
 	return (0);
 }
